MachineLearning/Model/ComputerVision/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
Flask-CORS==3.0.10
Werkzeug==2.3.7
mysql-connector-python==8.1.0
tensorflow==2.14.0
numpy==1.24.4